Split platformInfo
- Utility just loads it from env
- Main is the only place it is resolved fully
