+++ apps/studio/src-commercial/entrypoints/main.ts
@@ -10,6 +10,7 @@ import _ from 'lodash'
 // eslint-disable-next-line
 require('@electron/remote/main').initialize()
 log.initialize();
+log.transports.console.level = 'info'
 log.transports.file.level = "info"
 log.catchErrors({ showDialog: false})
 log.info("initializing background")
@@ -43,7 +44,7 @@ function initUserDirectory(d: string) {
 
 let utilityProcess: Electron.UtilityProcess
 // don't need this
-let newWindows: number[] = new Array();
+let newWindows: number[] = [];
 
 async function createUtilityProcess() {
   if (utilityProcess) {
@@ -51,13 +52,13 @@ async function createUtilityProcess() {
   }
 
   const args = {
-    isPackaged: `${electron.app.isPackaged}`,
-    locale: electron.app.getLocale(),
-    userDir: electron.app.getPath('userData'),
-    downloadDir: electron.app.getPath('downloads'),
-    homeDir: electron.app.getPath('home'),
-    shouldUseDarkColors: `${electron.nativeTheme.shouldUseDarkColors}`,
-    version: electron.app.getVersion()
+    // locale: electron.app.getLocale(),
+    // userDir: electron.app.getPath('userData'),
+    // downloadDir: electron.app.getPath('downloads'),
+    // homeDir: electron.app.getPath('home'),
+    // shouldUseDarkColors: `${electron.nativeTheme.shouldUseDarkColors}`,
+    // version: electron.app.getVersion(),
+    bksPlatformInfo: JSON.stringify(platformInfo)
   }
 
   utilityProcess = electron.utilityProcess.fork(
