+++ apps/studio/src/common/platform_info/mainPlatformInfo.ts
@@ -1,0 +1,89 @@
+
+import yargs from 'yargs-parser'
+import { resolve, join } from 'path'
+import { IPlatformInfo } from '../IPlatformInfo'
+
+// TODO: Automatically enable wayland without flags once
+// we're confident it will 'just work' for all Wayland users.
+const p = process
+
+export function mainPlatformInfo(): IPlatformInfo {
+
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+  const e = require('electron')
+  const platform = p.env.OS_OVERRIDE ? p.env.OS_OVERRIDE : p.platform
+  const testMode = p.env.TEST_MODE ? true : false
+  const isDevEnv = !e.app.isPackaged;
+  const isWindows = platform === 'win32'
+  const isMac = platform === 'darwin'
+  const isArm = p.arch.startsWith('arm')
+  const easyPlatform = isWindows ? 'windows' : (isMac ? 'mac' : 'linux')
+  const locale = e.app.getLocale();
+
+  const windowPrefersDarkMode = false
+
+  const updatesDisabled = !!p.env.BEEKEEPER_DISABLE_UPDATES
+
+  const oracleSupported = isMac && isArm ? false : true
+
+  const resourcesPath = isDevEnv ? resolve('./extra_resources') : resolve(p.resourcesPath)
+  let userDirectory = testMode ? './tmp' : e.app.getPath("userData")
+  const downloadsDirectory = testMode ? './tmp' : e.app.getPath('downloads')
+  const homeDirectory = testMode ? './tmp' : e.app.getPath('home')
+  if (p.env.PORTABLE_EXECUTABLE_DIR) {
+    userDirectory = join(p.env.PORTABLE_EXECUTABLE_DIR, 'beekeeper_studio_data')
+  }
+
+  const sessionType = p.env.XDG_SESSION_TYPE
+
+  const slice = isDevEnv ? 2 : 1
+  const parsedArgs = yargs(p.argv.slice(slice))
+
+
+
+  const appVersion = testMode ? 'test-mode' : e.app.getVersion()
+  const [major, minor, patch] = appVersion.split('.')
+  const parsedAppVersion = { major: Number(major), minor: Number(minor), patch: Number(patch) }
+
+  function isWaylandMode() {
+    return parsedArgs['ozone-platform-hint'] === 'auto' &&
+      sessionType === 'wayland' && !isWindows && !isMac
+  }
+   return {
+    isWindows, isMac, isArm, oracleSupported,
+    parsedArgs,
+    isLinux: !isWindows && !isMac,
+    sessionType,
+    isWayland: isWaylandMode(),
+    isSnap: p.env.ELECTRON_SNAP,
+    isPortable: isWindows && p.env.PORTABLE_EXECUTABLE_DIR,
+    isDevelopment: isDevEnv,
+    isAppImage: p.env.DESKTOPINTEGRATION === 'AppImageLauncher',
+    sshAuthSock: p.env.SSH_AUTH_SOCK,
+    environment: p.env.NODE_ENV,
+    resourcesPath,
+    env: {
+      development: isDevEnv,
+      test: testMode,
+      production: !isDevEnv && !testMode
+    },
+    debugEnabled: !!p.env.DEBUG,
+    DEBUG: p.env.DEBUG,
+    platform: easyPlatform,
+    darkMode: testMode ? true : !!(e.nativeTheme.shouldUseDarkColors) || windowPrefersDarkMode,
+    userDirectory,
+    downloadsDirectory,
+    homeDirectory,
+    testMode,
+    appDbPath: join(userDirectory, isDevEnv ? 'app-dev.db' : 'app.db'),
+    updatesDisabled,
+    appVersion,
+    parsedAppVersion,
+    // cloudUrl: isDevEnv ? 'https://staging.beekeeperstudio.io' : 'https://app.beekeeperstudio.io',
+    cloudUrl: 'https://app.beekeeperstudio.io',
+    locale,
+
+    // cloudUrl: isDevEnv ? 'http://localhost:3000' : 'https://app.beekeeperstudio.io'
+  }
+}
+
