+++ apps/studio/src/common/platform_info/utilityPlatformInfo.ts
@@ -1,0 +1,11 @@
+import { IPlatformInfo } from "../IPlatformInfo";
+import rawLog from 'electron-log'
+
+const log = rawLog.scope('utilityPlatformInfo')
+
+// why build it again from stratch? We just get it from the environment, thanks main process!
+export function utilityPlatformInfo(): IPlatformInfo {
+  const result = JSON.parse(process.env.bksPlatformInfo)
+  log.info(result)
+  return result
+}
