+++ apps/studio/src/lib/cloud/CloudClient.ts
@@ -29,13 +29,10 @@ const snakeCaseData: AxiosRequestTransformer = (data) => {
 
 const camelCaseData: AxiosResponseTransformer = (data) => {
   if (_.isPlainObject(data)) {
-    console.log('camel yes')
     const result = _.deepMapKeys(data, (_value, key) => _.camelCase(key))
-    log.info('camel result', result)
     return result
 
   }
-  console.log('camel no')
   return data
 }
